Add support for slide decks
Now we can add decks to use during class. These run in-browser and can 
be accessed by students during class to help take notes, or after class 
for reference or review.

For now, an initial deck has been added on Screen Typography.

This should be considered an experimental starting point for this 
feature.